WIP: added more edge-cases for filtering
When you add a filter for user-filters you need to supply
the user-type as contact or user. It uses any by default
which takes the path to query .

When you add the group-filters you need to set the type to group
so that the path to query groups can be triggered.

When using people-filters you just use it as is.