fix: prevent unsafe access to dom nodes in getDocumentDirection (#1788)
If a provider is in the head using mgt-loader the synchronous loading of mgt scripts
causes LocaliztionHelper to attempt to access the body DOM node of the document before
the browser has parsed the entire document causing body to be null. Using optional chaining
avoids trying to call a method on a null object.