fix: Redefine the focus on the input
Moving the _isFocused value to be checked after loading a the people so that we avoid having
empty people on the first call to loadState() or when you focus with the keyboard or a click.

Note: focus events do not bubble upwards. Not sure what to make of the focus on the input
element because the click event will be bubbled up and trigger the same function it calls
in the div encapsulating it.