feat: set the title text according to available parameters
If there are initials, set the title to indicate they are initials. If the image is availabele set the title with the 'photo for' prefix. If we have email set the title with the 'email address' prefix. Empty titles will be set as empty strings. This will throw accessibility bugs when you run the AI tool [aria-label text cannot be empty]