fix: update the text announced for person image, initials and email address (#1923)
Closes #1907

* chore: remove unused variable

* fix: change the alt text for person image

The new alt text includes 'photo for <person name>' prefix which differentiates the image and name announcements of a person component

* fix: change the alt text for person image

The new alt text includes 'photo for <person name>' prefix which differentiates the image and name announcements of a person component

* feat: add localization strings for person component

* feat: set the title text according to available parameters

If there are initials, set the title to indicate they are initials. If the image is availabele set the title with the 'photo for' prefix. If we have email set the title with the 'email address' prefix. Empty titles will be set as empty strings. This will throw accessibility bugs when you run the AI tool [aria-label text cannot be empty]

* fix: set the license text on all files