fix: set the person details in all blocks (#1959)
When you try to check for the id from  only, the value will be  in the event  were set on the component. To mitigate getting the default presence value, I update the instances that the  method updates the person's details. This ensures the final call to get the presence has either a user ID to make the call, calls the  endpoint or falls to the default presence value

Closes #1593

Co-authored-by: Gavin Barron <gavinbarron@microsoft.com>