feat!: split static config and permission methods out of mgt-personon-card
splits out the logic for cacluating the scopes required for the mgt-person-card into a separate function and file
splits the static config for MgtPersonCard into a separate class
changes the schema of the config for MgtPersonCard
fixes an issue where setting isSendMessageVisible to false did not stop the quick message ui from rendering
adds tests to validate the permission sets

BREAKING CHANGE: MgtPersonCard no longer has a static config property.This config has been moved to the MgtPersonCardConfig class to allow developers to import the config and associated getMgtPersonCardScopes function at the top level of their applicaiton without automatically adding the weight of the full mgt-person-card component and dependencies to the entry file for their applications.
BREAKING CHANGE: The organization section for configuring MgtPersonCard may no longer be set to false, use undefined instead.