feat: scope aware requests (#2797)
Adds helpers to determine if current token has a required scope
Adds helper to calculate the minimum scope to be requested if necessary
Fix test for prepScopes to match new behavior

BREAKING CHANGE: prepScopes now accepts an array of scopes, this is checked against the set of currently consented scopes for the user. If any of the supplied scopes are found, then no additional scopes are requested. If no match is found then the user will be prompted to consent to the first scope in the supplied array of scopes

fix: getGroupImage now correctly requires group.read.all or group.readwrite.all

BREAKING CHANGE: applications using mgt-person with fetch-image and person-detal where the supplied value is a group will now need to consent to either Group.Read.All or Group.ReadWrite.All. This replaces the existing behavior where the group image would silently fail to load and show an http 403 error in the console

fix: updated todo and planner permissions

BREAKING CHANGE: minimal permission for planner calls changed from Group.ReadWrite.All to Tasks.ReadWrite for write operation and from Group.Read.All to Tasks.Read for read operations