Update the StatefulGraphChatClient to set the notificationClient
Setting the client calls an external function graph that throws a TypeError. The error is thrown
when accessing the forComponent method. The assumption is that the Provider hasn't set up by the time
the StatefulGraphChatClient is instatiated,

This change leverages the Provider onStateChange events and updates the notificationClient appropriately
after the constructor initialization.